+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+[Unreleased]
+---------------------
+##### Removed
+- function `printForm` in print script **WARNING: Rewrite function call and use the `print` function as an example**.
+
 [4.48.1] - 2018-04-06
 ---------------------
 ##### Fixed